requirements.txt
torch
torchvision
numpy>=1.22.0
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pillow>=9.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
seaborn>=0.12.0 